backend/requirements.txt
fastapi
pydantic
psycopg2-binary
mysql-connector-python
starlette
azure-identity
typing
azure-storage-blob
unstructured[pdf]
langchain
langchain_community
langchain_core
langchain_openai
pydantic
python-dotenv
SQLAlchemy
uvicorn
opencv-python
sentence-transformers
chromadb
psutil
